requirements.txt
botbuilder-core>=4.7.0
aiohttp
openai
numpy
pandas
matplotlib
tiktoken
langchain
faiss-cpu
